:recycle: : Remove data providers from main target
