+++ Targets/MorningBearDataProvider/Sources/MyBadgeDataProvider.swift
@@ -7,3 +7,4 @@
 //
 
 import Foundation
+
